conduit-axum: Extract `response` module
